Fix mirroring
Previously mirroring actually changed the number of bytes generated,
wrapping them on the required axes. This meant that mirroring didn't
actually result in a mirrored version of the original, and instead
caused shifts in the image instead.

This commit changes the code to generate an entire image worth of bytes
and then apply mirroring whilst drawing the image to the array.

Some additional changes were made to simplify the code - primarily
dropping the width and height variables used to dictate the image size
(it's always 8x8).